`Refactored Security component to display password last changes, login notifications, and recent account activity.`
